% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherence.R
\name{choose_treatment}
\alias{choose_treatment}
\title{The choice function implied by categorical monotonicity}
\usage{
choose_treatment(z, a)
}
\arguments{
\item{z}{Character vector of all treatment alternatives, sorted in
\emph{decreasing} order of encouragement by the instrument. In the pricing
application this is the alternatives ordered from cheapest to most
expensive.}

\item{a}{Character vector of the alternatives the decision team adheres to,
i.e. one adherence set.}
}
\value{
A length-one character vector: the element of \code{a} appearing earliest
in \code{z}.
}
\description{
Categorical monotonicity states that the treatment received is a
deterministic function of the instrument and the adherence set: the decision
team picks the alternative in the adherence set that the instrument
encourages most.
}
\examples{
choose_treatment(c("a", "b", "c"), c("b", "c"))

}
